labs/lab12/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lab12"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
lab12-cli = "lab12.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lab12"]
